+++ lib/gibber/gibber-muc-connection.h
@@ -107,8 +107,8 @@ gibber_muc_connection_send(GibberMucConnection *connection,
                           GibberXmppStanza *stanza, GError **error);
 
 gboolean
-gibber_muc_connection_send_raw(GibberMucConnection *connection, 
-                               guint8 stream_id, const guint8 *data, 
+gibber_muc_connection_send_raw(GibberMucConnection *connection,
+                               guint16 stream_id, const guint8 *data,
                                gsize size, GError **error);
 
 G_END_DECLS
