+++ lib/gibber/gibber-r-multicast-transport.c
@@ -1306,7 +1306,7 @@ gibber_r_multicast_transport_connect(GibberRMulticastTransport *transport,
 
 gboolean
 gibber_r_multicast_transport_send(GibberRMulticastTransport *transport,
-                                  guint8 stream_id,
+                                  guint16 stream_id,
                                   const guint8 *data,
                                   gsize size,
                                   GError **error) {
